fix(trace): remove duplicate payload panel from queue picker
QueueEventList already renders its own two-column layout with a payload
panel. The SplitPane wrapper in trace-queue-picker was adding a second,
external payload panel — and its overflow-y-auto on the left pane was
preventing the inner grid columns from scrolling independently.

Remove the SplitPane + right-snippet payload panel from trace-queue-picker
and render QueueEventList directly. Move S3 link detection into
QueueEventList's own payload panel so the feature is preserved and shared
with the dashboard event viewer.

Task: ES-3029
